Migrate timepicker to styled-components
